+++ src/StorageBase/Action/WebHttpHandlerConfiguration/Set/WebHttpHandlerConfigurationSetPayload.php
@@ -4,11 +4,15 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\Base\Action\WebHttpHandlerConfiguration\Set;
 
+use Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface;
+use Heptacom\HeptaConnect\Dataset\Base\Support\AttachmentAwareTrait;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Create\CreatePayloadInterface;
 
-class WebHttpHandlerConfigurationSetPayload implements CreatePayloadInterface
+final class WebHttpHandlerConfigurationSetPayload implements CreatePayloadInterface, AttachmentAwareInterface
 {
+    use AttachmentAwareTrait;
+
     protected PortalNodeKeyInterface $portalNodeKey;
 
     protected string $path;